+++ models/Strategy.py
@@ -51,7 +51,9 @@ class Strategy:
                 > (self._df["close"].max() * (1 - self.app.noBuyNearHighPcnt() / 100))
             )
         ):
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 log_text = (
                     str(now)
                     + " | "
@@ -71,14 +73,14 @@ class Strategy:
             return False
 
         ## if last_action was set to "WAIT" due to an API problem, do not buy
-        if ( self.state.last_action == "WAIT"):
-            return False
-            log_text = (f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | last_action is WAIT, do not buy yet")
+        if self.state.last_action == "WAIT":
+            log_text = f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | last_action is WAIT, do not buy yet"
             Logger.warning(log_text)
+            return False
 
         # if EMA, MACD are disabled, do not buy
         if self.app.disableBuyEMA() and self.app.disableBuyMACD():
-            log_text = (f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | EMA, MACD indicators are disabled")
+            log_text = f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | EMA, MACD indicators are disabled"
             Logger.warning(log_text)
 
             return False
@@ -189,31 +191,44 @@ class Strategy:
         if debug:
             Logger.warning("\n*** isSellTrigger ***\n")
             Logger.warning("-- ignoring sell signal --")
-            Logger.warning(f"self.app.nosellminpcnt is None (nosellminpcnt: {self.app.nosellminpcnt})")
+            Logger.warning(
+                f"self.app.nosellminpcnt is None (nosellminpcnt: {self.app.nosellminpcnt})"
+            )
             Logger.warning(f"margin >= self.app.nosellminpcnt (margin: {margin})")
-            Logger.warning(f"margin <= self.app.nosellmaxpcnt (nosellmaxpcnt: {self.app.nosellmaxpcnt})")
+            Logger.warning(
+                f"margin <= self.app.nosellmaxpcnt (nosellmaxpcnt: {self.app.nosellmaxpcnt})"
+            )
             Logger.warning("\n")
 
         if (
-            ((self.app.nosellminpcnt is not None)
-                and (margin >= self.app.nosellminpcnt))
-                and ((self.app.nosellmaxpcnt is not None)
-                and (margin <= self.app.nosellmaxpcnt)
-            )):
-                if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
-                    log_text = "! Ignore Sell Signal (Within No-Sell Bounds)"
-                    Logger.warning(log_text)
-                return False
+            (self.app.nosellminpcnt is not None) and (margin >= self.app.nosellminpcnt)
+        ) and (
+            (self.app.nosellmaxpcnt is not None) and (margin <= self.app.nosellmaxpcnt)
+        ):
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
+                log_text = "! Ignore Sell Signal (Within No-Sell Bounds)"
+                Logger.warning(log_text)
+            return False
 
         if debug:
             Logger.warning("\n*** isSellTrigger ***\n")
             Logger.warning("-- loss failsafe sell at fibonacci band --")
-            Logger.warning(f"self.app.disableFailsafeFibonacciLow() is False (actual: {self.app.disableFailsafeFibonacciLow()})")
-            Logger.warning(f"self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})")
-            Logger.warning(f"self.app.sellLowerPcnt() is None (actual: {self.app.sellLowerPcnt()})")
+            Logger.warning(
+                f"self.app.disableFailsafeFibonacciLow() is False (actual: {self.app.disableFailsafeFibonacciLow()})"
+            )
+            Logger.warning(
+                f"self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})"
+            )
+            Logger.warning(
+                f"self.app.sellLowerPcnt() is None (actual: {self.app.sellLowerPcnt()})"
+            )
             Logger.warning(f"self.state.fib_low {self.state.fib_low} > 0")
             Logger.warning(f"self.state.fib_low {self.state.fib_low} >= {float(price)}")
-            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning(
+                f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)"
+            )
             Logger.warning("\n")
 
         # loss failsafe sell at fibonacci band
@@ -225,17 +240,29 @@ class Strategy:
             and self.state.fib_low >= float(price)
             and (self.app.allowSellAtLoss() or margin > 0)
         ):
-            log_text = (f"! Loss Failsafe Triggered (Fibonacci Band: {str(self.state.fib_low)})")
+            log_text = (
+                f"! Loss Failsafe Triggered (Fibonacci Band: {str(self.state.fib_low)})"
+            )
             Logger.warning(log_text)
-            self.app.notifyTelegram(f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}")
+            self.app.notifyTelegram(
+                f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}"
+            )
             return True
 
         if debug:
             Logger.warning("-- loss failsafe sell at trailing_stop_loss --")
-            Logger.warning(f"self.app.trailingStopLoss() != None (actual: {self.app.trailingStopLoss()})")
-            Logger.warning(f"change_pcnt_high ({change_pcnt_high}) < self.app.trailingStopLoss() ({self.app.trailingStopLoss()})")
-            Logger.warning(f"margin ({margin}) > self.app.trailingStopLossTrigger() ({self.app.trailingStopLossTrigger()})")
-            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning(
+                f"self.app.trailingStopLoss() != None (actual: {self.app.trailingStopLoss()})"
+            )
+            Logger.warning(
+                f"change_pcnt_high ({change_pcnt_high}) < self.app.trailingStopLoss() ({self.app.trailingStopLoss()})"
+            )
+            Logger.warning(
+                f"margin ({margin}) > self.app.trailingStopLossTrigger() ({self.app.trailingStopLossTrigger()})"
+            )
+            Logger.warning(
+                f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)"
+            )
             Logger.warning("\n")
 
         # loss failsafe sell at trailing_stop_loss
@@ -243,22 +270,37 @@ class Strategy:
             self.app.trailingStopLoss() != None
             and change_pcnt_high < self.app.trailingStopLoss()
             and margin > self.app.trailingStopLossTrigger()
-            and (self.app.allowSellAtLoss() or margin > 0)):
+            and (self.app.allowSellAtLoss() or margin > 0)
+        ):
 
-            log_text = (f"! Trailing Stop Loss Triggered (< {str(self.app.trailingStopLoss())}%)")
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            log_text = f"! Trailing Stop Loss Triggered (< {str(self.app.trailingStopLoss())}%)"
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 Logger.warning(log_text)
 
-            self.app.notifyTelegram(f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}")
+            self.app.notifyTelegram(
+                f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}"
+            )
             return True
 
         if debug:
             Logger.warning("-- loss failsafe sell at sell_lower_pcnt --")
-            Logger.warning(f"self.app.disableFailsafeLowerPcnt() is False (actual: {self.app.disableFailsafeLowerPcnt()})")
-            Logger.warning(f"and self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})")
-            Logger.warning(f"and self.app.sellLowerPcnt() != None (actual: {self.app.sellLowerPcnt()})")
-            Logger.warning(f"and margin ({margin}) < self.app.sellLowerPcnt() ({self.app.sellLowerPcnt()})")
-            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning(
+                f"self.app.disableFailsafeLowerPcnt() is False (actual: {self.app.disableFailsafeLowerPcnt()})"
+            )
+            Logger.warning(
+                f"and self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})"
+            )
+            Logger.warning(
+                f"and self.app.sellLowerPcnt() != None (actual: {self.app.sellLowerPcnt()})"
+            )
+            Logger.warning(
+                f"and margin ({margin}) < self.app.sellLowerPcnt() ({self.app.sellLowerPcnt()})"
+            )
+            Logger.warning(
+                f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)"
+            )
             Logger.warning("\n")
 
         # loss failsafe sell at sell_lower_pcnt
@@ -269,9 +311,13 @@ class Strategy:
             and margin < self.app.sellLowerPcnt()
             and (self.app.allowSellAtLoss() or margin > 0)
         ):
-            log_text = ("! Loss Failsafe Triggered (< " + str(self.app.sellLowerPcnt()) + "%)")
+            log_text = (
+                "! Loss Failsafe Triggered (< " + str(self.app.sellLowerPcnt()) + "%)"
+            )
             Logger.warning(log_text)
-            self.app.notifyTelegram(f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}")
+            self.app.notifyTelegram(
+                f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}"
+            )
             return True
 
         if debug:
@@ -297,12 +343,14 @@ class Strategy:
             and margin > self.app.sellUpperPcnt()
             and (self.app.allowSellAtLoss() or margin > 0)
         ):
-            log_text = (
-                f"! Profit Bank Triggered (> {str(self.app.sellUpperPcnt())}%)"
-            )
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            log_text = f"! Profit Bank Triggered (> {str(self.app.sellUpperPcnt())}%)"
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 Logger.warning(log_text)
-            self.app.notifyTelegram(f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}")
+            self.app.notifyTelegram(
+                f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}"
+            )
             return True
 
         if debug:
@@ -326,10 +374,14 @@ class Strategy:
             and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = "! Profit Bank Triggered (Selling At Resistance)"
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 Logger.warning(log_text)
             if not (not self.app.allowSellAtLoss() and margin <= 0):
-                self.app.notifyTelegram(f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}")
+                self.app.notifyTelegram(
+                    f"{self.app.getMarket()} ({self.app.printGranularity()}) {log_text}"
+                )
             return True
 
         return False
@@ -342,9 +394,13 @@ class Strategy:
             Logger.warning("\n*** isWaitTrigger ***\n")
 
         if debug and self.state.action == "BUY":
-            Logger.warning("-- if bear market and bull only return true to abort buy --")
+            Logger.warning(
+                "-- if bear market and bull only return true to abort buy --"
+            )
             Logger.warning(f"self.state.action == 'BUY' (actual: {self.state.action})")
-            Logger.warning(f"and self.app.disableBullOnly() is True (actual: {self.app.disableBullOnly()})")
+            Logger.warning(
+                f"and self.app.disableBullOnly() is True (actual: {self.app.disableBullOnly()})"
+            )
             Logger.warning(f"and goldencross is False (actual: {goldencross})")
             Logger.warning("\n")
 
@@ -361,7 +417,9 @@ class Strategy:
         if debug and self.state.action == "SELL":
             Logger.warning("-- configuration specifies to not sell at a loss --")
             Logger.warning(f"self.state.action == 'SELL' (actual: {self.state.action})")
-            Logger.warning(f"and self.app.allowSellAtLoss() is False (actual: {self.app.allowSellAtLoss()})")
+            Logger.warning(
+                f"and self.app.allowSellAtLoss() is False (actual: {self.app.allowSellAtLoss()})"
+            )
             Logger.warning(f"and margin ({margin}) <= 0")
             Logger.warning("\n")
 
@@ -371,7 +429,9 @@ class Strategy:
             and not self.app.allowSellAtLoss()
             and margin <= 0
         ):
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 log_text = "! Ignore Sell Signal (No Sell At Loss)"
                 Logger.warning(log_text)
             return True
@@ -401,7 +461,9 @@ class Strategy:
                 and (margin <= self.app.nosellmaxpcnt)
             )
         ):
-            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+            if not app.isSimulation() or (
+                app.isSimulation() and not app.simResultOnly()
+            ):
                 log_text = "! Ignore Sell Signal (Within No-Sell Bounds)"
                 Logger.warning(log_text)
             return True